Refactor rate limit handling by moving RateLimitHandler to utils; update import paths accordingly
